Refactor of the fitter
