+++ EasyReflectometryApp/Gui/Components/AnalysisFitables.qml
@@ -12,7 +12,7 @@ import Gui.Globals 1.0 as ExGlobals
 EaComponents.TableView {
     id: table
 
-    enabled: ExGlobals.Constants.proxy.isFitFinished
+    enabled: ExGlobals.Constants.proxy.fitter.isFitFinished
 
     maxRowCountShow: 8
     defaultInfoText: qsTr("No Parameters Found")
