+++ EasyReflectometryApp/Gui/Components/ResultsDialog.qml
@@ -10,8 +10,8 @@ import Gui.Globals 1.0 as ExGlobals
 EaElements.Dialog {
     id: dialog
 
-    property bool gotResults: typeof ExGlobals.Constants.proxy.fitResults.nvarys !== 'undefined' &&
-                              ExGlobals.Constants.proxy.isFitFinished
+    property bool gotResults: typeof ExGlobals.Constants.proxy.fitter.fitResults.nvarys !== 'undefined' &&
+                              ExGlobals.Constants.proxy.fitter.isFitFinished
 
     title: qsTr("Refinement Results")
 
@@ -22,21 +22,21 @@ EaElements.Dialog {
     Column {
         EaElements.Label {
             text: gotResults
-                  ? `Success: ${ExGlobals.Constants.proxy.fitResults.success}`
+                  ? `Success: ${ExGlobals.Constants.proxy.fitter.fitResults.success}`
                   : `Fitting cancelled`
         }
 
         EaElements.Label {
             enabled: gotResults
             text: gotResults
-                  ? `Num. refined parameters: ${ExGlobals.Constants.proxy.fitResults.nvarys}`
+                  ? `Num. refined parameters: ${ExGlobals.Constants.proxy.fitter.fitResults.nvarys}`
                   : ""
         }
 
         EaElements.Label {
             enabled: gotResults
             text: gotResults
-                  ? `Goodness-of-fit (reduced \u03c7\u00b2): ${ExGlobals.Constants.proxy.fitResults.redchi2.toFixed(2)}`
+                  ? `Goodness-of-fit (reduced \u03c7\u00b2): ${ExGlobals.Constants.proxy.fitter.fitResults.redchi2.toFixed(2)}`
                   : ""
         }
     }
